requirements.txt
requests
beautifulsoup4
lxml
aiogram
python-decouple